refactor: move tests files to test dir
