+++ tsconfig.build.json
@@ -3,7 +3,7 @@
   "compilerOptions": {
     "sourceMap": false,
   },
-  "exclude": [
-    "src/**/*.spec.ts"
+  "include": [
+    "src/**/*.ts"
   ],
 }
